+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds prop validation for the `TextInput` "number" type for the `min` and `max` props.
+
 ### Updates
 
 - Updates the `TemplateAppContainer`'s responsive styles for its sidebar component.